replace : with shared toast helper
