+++ tests/e2e/utils/page-objects/savedRepliesPage.ts
@@ -1,5 +1,6 @@
 import { expect, Locator, Page } from "@playwright/test";
 import { BasePage } from "./basePage";
+import { waitForToast } from "../toastHelpers";
 
 export class SavedRepliesPage extends BasePage {
   readonly pageTitle: Locator;
@@ -213,31 +214,7 @@ export class SavedRepliesPage extends BasePage {
   }
 
   async waitForToast(message: string) {
-    try {
-      const toastSelectors = [
-        `[role="alert"]:has-text("${message}")`,
-        `[data-testid="toast"]:has-text("${message}")`,
-        `.toast:has-text("${message}")`,
-        `*:has-text("${message}")[role="status"]`,
-      ];
-
-      let toastFound = false;
-      for (const selector of toastSelectors) {
-        try {
-          await this.page.locator(selector).waitFor({ state: "visible", timeout: 2000 });
-          toastFound = true;
-          break;
-        } catch {
-          // Try next selector
-        }
-      }
-
-      if (!toastFound) {
-        await this.page.waitForSelector(`text="${message}"`, { timeout: 1000 });
-      }
-    } catch (error) {
-      console.log(`Toast message "${message}" not found, continuing...`);
-    }
+    await waitForToast(this.page, message);
   }
 
   async openCreateDialog() {
